backend/pyproject.toml
[project]
name = "signsense-backend"
version = "0.1.0"
description = "SignSense AI - Real-Time ASL Interpreter Backend"
requires-python = ">=3.12"
dependencies = [
    "vision-agents[getstream,gemini,elevenlabs,deepgram]",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.1.0",
    "inference-sdk>=0.9.0",
    "stream-chat>=4.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
